+++ lib/ruby_smart/simple_logger/scenes.rb
@@ -11,9 +11,8 @@ module RubySmart
         # > ================================================= [Debug] ================================================
         # > "DEBUGGED DATA" <- analyzed by awesome_print#ai method
         # > ==========================================================================================================
-        base.scene :debug, { level: :debug, inspect: true, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, *args|
-          subject, opts = _scene_subject_with_opts(args, 'Debug')
-          self.log data, _scene_opt(:debug, { subject: subject }, opts)
+        base.scene :debug, { level: :debug, inspect: true, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, subject = 'Debug', **opts|
+          self.log data, _scene_opts(:debug, subject: subject.to_s, **opts)
         end
 
         # info method (BASE)
@@ -23,9 +22,8 @@ module RubySmart
         # > ================================================= [Info] =================================================
         # > DATA
         # > ==========================================================================================================
-        base.scene :info, { level: :info, mask: { clr: :cyan }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, *args|
-          subject, opts = _scene_subject_with_opts(args, 'Info')
-          self.log data, _scene_opt(:info, { subject: subject }, opts)
+        base.scene :info, { level: :info, mask: { clr: :cyan }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, subject = 'Info', **opts|
+          self.log data, _scene_opts(:info, subject: subject.to_s, **opts)
         end
 
         # warn method (BASE)
@@ -35,9 +33,8 @@ module RubySmart
         # > ================================================= [Warn] =================================================
         # > DATA
         # > ==========================================================================================================
-        base.scene :warn, { level: :warn, mask: { clr: :yellow }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, *args|
-          subject, opts = _scene_subject_with_opts(args, 'Warn')
-          self.log data, _scene_opt(:warn, { subject: subject }, opts)
+        base.scene :warn, { level: :warn, mask: { clr: :yellow }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, subject = 'Warn', **opts|
+          self.log data, _scene_opts(:warn, subject: subject.to_s, **opts)
         end
 
         # error method (BASE)
@@ -47,9 +44,8 @@ module RubySmart
         # > ================================================ [Error] =================================================
         # > DATA
         # > ==========================================================================================================
-        base.scene :error, { level: :error, mask: { clr: :red }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, *args|
-          subject, opts = _scene_subject_with_opts(args, 'Error')
-          self.log data, _scene_opt(:error, { subject: subject }, opts)
+        base.scene :error, { level: :error, mask: { clr: :red }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, subject = 'Error', **opts|
+          self.log data, _scene_opts(:error, subject: subject.to_s, **opts)
         end
 
         # fatal method (BASE)
@@ -59,9 +55,8 @@ module RubySmart
         # > ================================================ [Fatal] =================================================
         # > DATA
         # > ==========================================================================================================
-        base.scene :fatal, { level: :fatal, mask: { clr: :bg_red }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, *args|
-          subject, opts = _scene_subject_with_opts(args, 'Fatal')
-          self.log data, _scene_opt(:fatal, { subject: subject }, opts)
+        base.scene :fatal, { level: :fatal, mask: { clr: :bg_red }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, subject = 'Fatal', **opts|
+          self.log data, _scene_opts(:fatal, subject: subject.to_s, **opts)
         end
 
         # unknown method (BASE)
@@ -71,9 +66,8 @@ module RubySmart
         # > =============================================== [Unknown] ================================================
         # > DATA
         # > ==========================================================================================================
-        base.scene :unknown, { level: :unknown, mask: { clr: :gray }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, *args|
-          subject, opts = _scene_subject_with_opts(args, 'Unknown')
-          self.log data, _scene_opt(:unknown, { subject: subject }, opts)
+        base.scene :unknown, { level: :unknown, mask: { clr: :gray }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, subject = 'Unknown', **opts|
+          self.log data, _scene_opts(:unknown, subject: subject.to_s, **opts)
         end
 
         # success method
@@ -83,9 +77,8 @@ module RubySmart
         # > ================================================ [Success] ================================================
         # > DATA
         # > ===========================================================================================================
-        base.scene :success, { level: :success, mask: { clr: :green }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, *args|
-          subject, opts = _scene_subject_with_opts(args, 'Success')
-          self.log data, _scene_opt(:success, { subject: subject }, opts)
+        base.scene :success, { level: :success, mask: { clr: :green }, payload: [[:mask, ' [%{subject}] '], :__data__, :mask] } do |data, subject = 'Success', **opts|
+          self.log data, _scene_opts(:success, subject: subject.to_s, **opts)
         end
 
         # header method
@@ -95,11 +88,11 @@ module RubySmart
         # > ===========================================================================================================
         # > ================================================ <Subject> ================================================
         # > ===========================================================================================================
-        base.scene :header, { level: :debug, payload: [:mask, [:mask, ' <%{subject}> '], :mask] } do |subject, opts = {}|
+        base.scene :header, { level: :debug, payload: [:mask, [:mask, ' <%{subject}> '], :mask] } do |subject, **opts|
           # autostart a timer method, if required
           self.timer(:start, :default) if opts[:timer]
 
-          self.log nil, _scene_opt(:header, { subject: subject }, opts)
+          self.log nil, _scene_opts(:header, subject: subject.to_s, **opts)
         end
 
         # footer method
@@ -109,8 +102,8 @@ module RubySmart
         # > ===========================================================================================================
         # > ================================================ >Subject< ================================================
         # > ===========================================================================================================
-        base.scene :footer, { level: :debug, payload: [:mask, [:mask, ' >%{subject}< '], :mask] } do |subject, opts = {}|
-          self.log nil, _scene_opt(:footer, { subject: subject }, opts)
+        base.scene :footer, { level: :debug, payload: [:mask, [:mask, ' >%{subject}< '], :mask] } do |subject, **opts|
+          self.log nil, _scene_opts(:footer, subject: subject.to_s, **opts)
 
           # clears & prints timer
           self.desc("duration: #{self.timer(:clear, :default, :humanized => true)}") if opts[:timer]
@@ -123,8 +116,8 @@ module RubySmart
         # > --------------------------------------------------------------------------------
         # > #----------------------------------- Subject ----------------------------------#
         # > --------------------------------------------------------------------------------
-        base.scene :topic, { level: :debug, mask: { char: '-', length: 95, clr: :blueish }, payload: [:mask, [:mask, '%{title}'], :mask] } do |subject, opts = {}|
-          opts         = _scene_opt(:topic, opts)
+        base.scene :topic, { level: :debug, mask: { char: '-', length: 95, clr: :blueish }, payload: [:mask, [:mask, '%{title}'], :mask] } do |subject, **opts|
+          opts         = _scene_opts(:topic, **opts)
           txt          = " #{subject} ".center(opts[:mask][:length] - 2, opts[:mask][:char])
           opts[:title] = "##{txt}#"
 
@@ -137,8 +130,8 @@ module RubySmart
         #
         # > # Subject
         # > ----------------------------------------------------------------------
-        base.scene :theme, { level: :debug, clr: :purple, mask: { char: '-', length: 85, clr: :purple }, payload: [[:txt, '# %{subject}'], :mask] } do |subject, opts = {}|
-          self.log nil, _scene_opt(:theme, { subject: subject }, opts)
+        base.scene :theme, { level: :debug, clr: :purple, mask: { char: '-', length: 85, clr: :purple }, payload: [[:txt, '# %{subject}'], :mask] } do |subject, **opts|
+          self.log nil, _scene_opts(:theme, subject: subject.to_s, **opts)
         end
 
         # theme_result method
@@ -148,9 +141,9 @@ module RubySmart
         # > ----------------------------------------------------------------------
         # > -> Result
         # >
-        base.scene :theme_result, { level: :debug, mask: { char: '-', length: 85, clr: :purple }, payload: [:mask, [:txt, '-> %{result}'], ''] } do |result, status = nil, opts = {}|
+        base.scene :theme_result, { level: :debug, mask: { char: '-', length: 85, clr: :purple }, payload: [:mask, [:txt, '-> %{result}'], ''] } do |result, status = nil, **opts|
           res_or_clr = status.nil? ? result : status
-          self.log nil, _scene_opt(:theme_result, { result: result, clr: _res_clr(res_or_clr) }, opts)
+          self.log nil, _scene_opts(:theme_result, result: result, clr: _res_clr(res_or_clr), **opts)
         end
 
         # theme_line method
@@ -158,8 +151,8 @@ module RubySmart
         # prints: colored line with no text
         #
         # > ----------------------------------------------------------------------
-        base.scene :theme_line, { level: :debug, mask: { char: '-', length: 85, clr: :purple }, payload: [:mask] } do |opts = {}|
-          self.log nil, _scene_opt(:theme_line, opts)
+        base.scene :theme_line, { level: :debug, mask: { char: '-', length: 85, clr: :purple }, payload: [:mask] } do |**opts|
+          self.log nil, _scene_opts(:theme_line, **opts)
         end
 
         # desc method
@@ -168,8 +161,8 @@ module RubySmart
         #
         # > "description"
         # >
-        base.scene :desc, { level: :debug, clr: :purple, payload: [[:txt, '%{description}']] } do |description, opts = {}|
-          self.log nil, _scene_opt(:desc, { description: description.to_s }, opts)
+        base.scene :desc, { level: :debug, clr: :purple, payload: [[:txt, '%{description}']] } do |description, **opts|
+          self.log nil, _scene_opts(:desc, description: description.to_s, **opts)
         end
 
         # job method
@@ -179,8 +172,8 @@ module RubySmart
         #
         # > - Job name                                                         =>
         #     ________________________________________________________________ <- 64 chars
-        base.scene :job, { level: :debug, clr: :cyan, nl: false, length: 64, payload: [[:concat, ['- ', [:txt, '%{name}'], ' => ']]] } do |name, opts = {}, &block|
-          self.log nil, _scene_opt(:job, { name: name }, opts)
+        base.scene :job, { level: :debug, clr: :cyan, nl: false, length: 64, payload: [[:concat, ['- ', [:txt, '%{name}'], ' => ']]] } do |name, **opts, &block|
+          self.log nil, _scene_opts(:job, name: name, **opts)
           self.result(*block.call) if block_given?
         end
 
@@ -191,8 +184,8 @@ module RubySmart
         #
         # >   * Subjob name                                                    =>
         #       ______________________________________________________________ <- 62 chars
-        base.scene :sub_job, { level: :debug, clr: :cyan, nl: false, length: 62, payload: [[:concat, ['  * ', [:txt, '%{name}'], ' => ']]] } do |name, opts = {}, &block|
-          self.log nil, _scene_opt(:sub_job, { name: name }, opts)
+        base.scene :sub_job, { level: :debug, clr: :cyan, nl: false, length: 62, payload: [[:concat, ['  * ', [:txt, '%{name}'], ' => ']]] } do |name, **opts, &block|
+          self.log nil, _scene_opts(:sub_job, name: name, **opts)
           self.result(*block.call) if block_given?
         end
 
@@ -201,9 +194,9 @@ module RubySmart
         # prints: colored result
         #
         # > Result
-        base.scene :result, { level: :debug, payload: [[:txt, '%{result}']] } do |result, status = nil, opts = {}|
+        base.scene :result, { level: :debug, payload: [[:txt, '%{result}']] } do |result, status = nil, **opts|
           res_or_clr = status.nil? ? result : status
-          self.log nil, _scene_opt(:result, { result: result, clr: _res_clr(res_or_clr) }, opts)
+          self.log nil, _scene_opts(:result, result: result, clr: _res_clr(res_or_clr), **opts)
         end
 
         # job_result method
@@ -211,9 +204,9 @@ module RubySmart
         # prints: job with combined colored result
         #
         # > - Job name                                                         => Result
-        base.scene :job_result, { level: :debug } do |name, result, status = nil, opts = {}|
-          self.job(name, opts)
-          self.result(result, status, opts)
+        base.scene :job_result, { level: :debug } do |name, result, status = nil, **opts|
+          self.job(name, **opts)
+          self.result(result, status, **opts)
         end
 
         # sub_job_result method
@@ -221,9 +214,9 @@ module RubySmart
         # prints: sub_job with combined colored result
         #
         # >   * Subjob name                                                    => Result
-        base.scene :sub_job_result, { level: :debug } do |name, result, status = nil, opts = {}|
-          self.sub_job(name, opts)
-          self.result(result, status, opts)
+        base.scene :sub_job_result, { level: :debug } do |name, result, status = nil, **opts|
+          self.sub_job(name, **opts)
+          self.result(result, status, **opts)
         end
 
         # line method
@@ -231,8 +224,8 @@ module RubySmart
         # prints: just a line with data
         #
         # > DATA
-        base.scene :line, { level: :debug } do |data, opts = {}|
-          self.log data, _scene_opt(:line, opts)
+        base.scene :line, { level: :debug } do |data, **opts|
+          self.log data, _scene_opts(:line, **opts)
         end
 
         # print method
@@ -240,8 +233,8 @@ module RubySmart
         # prints: prints data without a newline
         #
         # > DATA
-        base.scene :print, { level: :debug, nl: false } do |data, opts = {}|
-          self.log data, _scene_opt(:print, opts)
+        base.scene :print, { level: :debug, nl: false } do |data, **opts|
+          self.log data, _scene_opts(:print, **opts)
         end
 
         # nl method
@@ -250,8 +243,8 @@ module RubySmart
         #
         # >
         # >
-        base.scene :nl, { level: :debug } do |opts = {}|
-          self.log '', _scene_opt(:nl, opts)
+        base.scene :nl, { level: :debug } do |**opts|
+          self.log '', _scene_opts(:nl, **opts)
         end
 
         # spec method
@@ -263,7 +256,7 @@ module RubySmart
         # "other"   => ? (yellow)
         #
         # > .FFF...??...F....F...F..???....F...??
-        base.scene :spec, { level: :debug, nl: false, payload: [[:txt, '%{result}']] } do |status, opts = {}|
+        base.scene :spec, { level: :debug, nl: false, payload: [[:txt, '%{result}']] } do |status, **opts|
           result = if status.is_a?(TrueClass)
                      '.'
                    elsif status.is_a?(FalseClass)
@@ -272,7 +265,7 @@ module RubySmart
                      status = :yellow
                      '?'
                    end
-          self.log nil, _scene_opt(:spec, { result: result, clr: _res_clr(status) }, opts)
+          self.log nil, _scene_opts(:spec, result: result, clr: _res_clr(status), **opts)
         end
 
         # progress method
@@ -283,7 +276,7 @@ module RubySmart
         # > - Progress of Step 1                               [ 40%] ===================>------------------------------
         #     ________________________________________________ <- 48 chars
         #                                                 50 chars -> __________________________________________________
-        base.scene :progress, { level: :debug, payload: [[:txt, '- %{name} [%{perc}%] %{progress}']] } do |name, perc, opts = {}|
+        base.scene :progress, { level: :debug, payload: [[:txt, '- %{name} [%{perc}%] %{progress}']] } do |name, perc, **opts|
           pmask_length = 50
 
           # convert and fix progress
@@ -298,7 +291,7 @@ module RubySmart
 
           progress_string = _clr(('=' * pmask_left_length) + '>', :green) + _clr('-' * pmask_right_length, :red)
           perc_string     = perc.to_s.rjust(3, ' ')
-          self.log nil, _scene_opt(:progress, { name: _clr(_lgth(name, 48), :cyan), perc: perc_string, progress: progress_string }, opts)
+          self.log nil, _scene_opts(:progress, name: _clr(_lgth(name, 48), :cyan), perc: perc_string, progress: progress_string, **opts)
         end
 
         # processed method
@@ -309,7 +302,7 @@ module RubySmart
         # ╟ doing some cool log
         # ╟ doing some extra log
         # ╚ END   ❯ job [SUCCESS] (duration: 4.34223)
-        base.scene :processed, { level: :debug } do |name, opts = {}, &block|
+        base.scene :processed, { level: :debug } do |name, **opts, &block|
           # increase level
           lvl = processed_lvl(:up)
 
@@ -326,7 +319,7 @@ module RubySmart
             self.timer(:start, timer_key)
 
             # send START name as +data+ - the full log line is created through the +_pcd+ method.
-            self.log(name, _scene_opt(:processed, opts, { pcd: :start }))
+            self.log(name, _scene_opts(:processed, **opts, pcd: :start))
 
             # run the provided block and resolve result
             result_str = case block.call
@@ -345,7 +338,7 @@ module RubySmart
             result_str ||= ''
 
             # send END name with result & possible time as +data+ - the full log line is created through the +_pcd+ method.
-            self.log("#{name} #{result_str}#{(timer_key ? "(#{self.timer(:clear, timer_key, humanized: true)})" : '')}", _scene_opt(:processed, opts, { pcd: :end }))
+            self.log("#{name} #{result_str}#{(timer_key ? "(#{self.timer(:clear, timer_key, humanized: true)})" : '')}", _scene_opts(:processed, **opts, pcd: :end ))
 
             # reduce level
             processed_lvl(:down)
@@ -357,26 +350,27 @@ module RubySmart
         # model method
         # log level @ error/success/info
         # prints: ActiveRecord::Base related data, depending on the models "save" state (also shows possible errors)
-        base.scene :model do |model, opts = {}|
+        base.scene :model do |model, status = nil, **opts|
           # build model-logging string
           mdl_string = "#{model.id.present? ? "##{model.id} - " : ''}#{model.to_s[0..49]}"
 
           # resolve model's status
-          status = ((model.persisted? && model.errors.empty?) ? (model.previous_changes.blank? ? :skipped : :success) : :error)
+          status     = ((model.persisted? && model.errors.empty?) ? (model.previous_changes.blank? ? :skipped : :success) : :error) if status.nil?
 
           # switch between status
           case status
           when :success
             # show verbose logging for updated records
             if opts[:verbose] != false && !model.previously_new_record?
-              log(:success, "#{mdl_string} (#{model.previous_changes.inspect})", tag: "#{model.class.name.upcase}|UPDATED")
+              self.success("#{mdl_string} (#{model.previous_changes.inspect})", tag: "#{model.class.name.upcase}|UPDATED", **opts)
             else
-              log(:success, mdl_string, tag: "#{model.class.name.upcase}|#{(model.previously_new_record? ? 'CREATED' : 'UPDATED')}")
+              self.success(mdl_string, tag: "#{model.class.name.upcase}|#{(model.previously_new_record? ? 'CREATED' : 'UPDATED')}", **opts)
             end
           when :error
-            log(:error, "#{mdl_string} (#{model.errors.full_messages.join(', ').presence || '-'})", tag: "#{model.class.name.upcase}|ERROR")
+            msg_string = model.errors.full_messages.join(', ')
+            self.error("#{mdl_string} (#{msg_string.present? ? msg_string : '-'})", tag: "#{model.class.name.upcase}|ERROR", **opts)
           else
-            log(:info, mdl_string, tag: "#{model.class.name.upcase}|#{status}")
+            self.info(mdl_string, tag: "#{model.class.name.upcase}|#{status}", **opts)
           end
         end
       end